Deployment/app/requirements.txt
flask==1.1.2
google-cloud-storage==1.31.0
google-cloud-bigquery==1.27.2
pandas==0.25.3